added sign up logic and refactored user

+++ pkg/repository/UserRepo.go
@@ -23,33 +23,42 @@ func (database *UserRepo) GetAll() ([]models.User, error) {
 func (database *UserRepo) GetList(id int) (models.User, error) {
 	var data models.User
 
-	database.db.Raw("SELECT * FROM users WHERE id = ?", id).Find(&data)
+	database.db.Preload("Ad").Raw("SELECT * FROM users WHERE id = ?", id).Find(&data)
 
 	return data, nil
 }
 
-func (database *UserRepo) Delete(id int) error {
+func (database *UserRepo) Delete(id int) (models.User, error) {
 	var data models.User
 
-	database.db.Raw("SELECT * FROM users WHERE id = ?", id).Find(&data)
+	err := database.db.Raw("SELECT * FROM users WHERE id = ?", id).Find(&data)
+
+	if data.Id == 0 {
+		return data, err.Error
+	}
 
 	database.db.Delete(&data)
 
-	return nil
+	return data, nil
 }
 
-func (database *UserRepo) Update(id int, data models.User) error {
+func (database *UserRepo) Update(id int, data models.User) (models.User, error) {
 	user := models.User{
 		Name:     data.Name,
 		Lastname: data.Lastname,
 		Username: data.Username,
 		Email:    data.Email,
 		Password: data.Password,
+		Ad:       data.Ad,
 	}
 
 	database.db.Model(&user).Where("id = ?", id).Updates(&user).Find(&user)
 
+	if user.Id == 0 {
+		return user, nil
+	}
+
 	database.db.Save(&user)
 
-	return nil
+	return user, nil
 }
